+++ packages/rest-api/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.8](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.8.7...@synapsecns/rest-api@1.8.8) (2024-11-19)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
+## [1.8.7](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.8.6...@synapsecns/rest-api@1.8.7) (2024-11-11)
+
+**Note:** Version bump only for package @synapsecns/rest-api
+
+
+
+
+
 ## [1.8.6](https://github.com/synapsecns/sanguine/compare/@synapsecns/rest-api@1.8.5...@synapsecns/rest-api@1.8.6) (2024-11-07)
 
 **Note:** Version bump only for package @synapsecns/rest-api