+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.40.19](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.18...@synapsecns/synapse-interface@0.40.19) (2024-11-19)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.40.18](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.17...@synapsecns/synapse-interface@0.40.18) (2024-11-12)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.40.17](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.40.16...@synapsecns/synapse-interface@0.40.17) (2024-11-07)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface